+++ RELEASES.md
@@ -3,6 +3,7 @@
 ## 0.8.3dev
 
 #### New features
+- Added feature to (Fused) Gromov-Wasserstein solvers herited from `ot.optim` to support relative and absolute loss variations as stopping criterions (PR #431)
 - Added feature to (Fused) Gromov-Wasserstein solvers to handle asymmetric matrices (PR #431)
 - Added semi-relaxed (Fused) Gromov-Wasserstein solvers in `ot.gromov` + examples (PR #431)
 - Added the spherical sliced-Wasserstein discrepancy in `ot.sliced.sliced_wasserstein_sphere` and `ot.sliced.sliced_wasserstein_sphere_unif` + examples (PR #434)